feat(auth): add me endpoint on backend

+++ backend/internal/store/store.go
@@ -10,7 +10,7 @@ var ErrNotFound = errors.New("not found")
 
 type UserStore interface {
 	Create(ctx context.Context, input model.CreateUserInput) error
-	GetByID(ctx context.Context, id int) (*model.User, error)
+	GetByID(ctx context.Context, id int64) (*model.User, error)
 	GetByEmail(ctx context.Context, email string) (*model.User, error)
 	EmailExists(ctx context.Context, email string) (bool, error)
 }
